requirements.txt
PySide6
Pillow
piexif
mutagen
PyPDF2
python-docx
python-pptx
openpyxl
opencv-python
pyinstaller